Rename MS Graph and Open API plugin examples
