+++ samples/dotnet/MsGraphPluginsExample/MsGraphPluginsExample.csproj
