+++ samples/dotnet/MsGraphPluginsExample/appsettings.json
