+++ samples/dotnet/OpenApiPluginsExample/GitHubPlugin/Model/User.cs
@@ -2,7 +2,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace GitHubSkill.Model;
+namespace GitHubPlugin.Model;
 
 /// <summary>
 /// Represents a user on GitHub.
