+++ samples/dotnet/OpenApiPluginsExample/GitHubPlugin/openapi.json
