+++ samples/dotnet/OpenApiPluginsExample/Program.cs
@@ -7,7 +7,7 @@ using System.Reflection;
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Threading.Tasks;
-using GitHubSkill.Model;
+using GitHubPlugin.Model;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.SemanticKernel;
@@ -18,18 +18,18 @@ using Microsoft.SemanticKernel.Planning;
 using Microsoft.SemanticKernel.Skills.OpenAPI.Authentication;
 using Microsoft.SemanticKernel.Skills.OpenAPI.Extensions;
 
-namespace OpenApiSkillsExample;
+namespace OpenApiPluginsExample;
 
 /// <summary>
-/// The chat example below is meant to demonstrate the use of an OpenAPI-based skill (e.g., GitHub),
+/// The chat example below is meant to demonstrate the use of an OpenAPI-based plugin (e.g., GitHub),
 /// a planner (e.g., ActionPlanner) and chat completions to create a conversational experience with
-/// additional information from a skill when needed.
+/// additional information from a plugin when needed.
 /// </summary>
 internal sealed class Program
 {
     private static async Task Main(string[] args)
     {
-        Console.WriteLine("Welcome to Semantic Kernel OpenAPI Skills Example!");
+        Console.WriteLine("Welcome to Semantic Kernel OpenAPI Plugins Example!");
 
         // Load configuration
         IConfigurationRoot configuration = new ConfigurationBuilder()
@@ -68,18 +68,18 @@ internal sealed class Program
 
         var kernel = kb.Build();
 
-        // Register the GitHub skill using an OpenAPI definition containing only pull request GET operations.
-        GitHubSkillOptions gitHubOptions = configuration.GetRequiredSection(GitHubSkillOptions.PropertyName).Get<GitHubSkillOptions>()
-            ?? throw new InvalidOperationException($"Missing configuration for {GitHubSkillOptions.PropertyName}.");
+        // Register the GitHub plugin using an OpenAPI definition containing only pull request GET operations.
+        GitHubPluginOptions gitHubOptions = configuration.GetRequiredSection(GitHubPluginOptions.PropertyName).Get<GitHubPluginOptions>()
+            ?? throw new InvalidOperationException($"Missing configuration for {GitHubPluginOptions.PropertyName}.");
 
         BearerAuthenticationProvider authenticationProvider = new(() => Task.FromResult(gitHubOptions.Key));
 
         await kernel.ImportAIPluginAsync(
-            skillName: "GitHubSkill",
-            filePath: Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)!, "GitHubSkill/openapi.json"),
+            skillName: "GitHubPlugin",
+            filePath: Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)!, "GitHubPlugin/openapi.json"),
             new OpenApiSkillExecutionParameters(authCallback: authenticationProvider.AuthenticateRequestAsync));
 
-        // Use a planner to decide when to call the GitHub skill. Since we are not chaining operations, use
+        // Use a planner to decide when to call the GitHub plugin. Since we are not chaining operations, use
         // the ActionPlanner which is a simplified planner that will always return a 0 or 1 step plan.
         ActionPlanner planner = new(kernel, logger: logger);
 
@@ -99,7 +99,7 @@ internal sealed class Program
 
             // Add GitHub's response, if any, to the chat history.
             int planResultTokenAllowance = (int)(aiOptions.TokenLimit * 0.25); // Allow up to 25% of our token limit to be from GitHub.
-            string planResult = await PlanGitHubSkillAsync(gitHubOptions, planner, chatHistory, input, planResultTokenAllowance, logger);
+            string planResult = await PlanGitHubPluginAsync(gitHubOptions, planner, chatHistory, input, planResultTokenAllowance, logger);
             if (!string.IsNullOrWhiteSpace(planResult))
             {
                 chatHistory.AddUserMessage(planResult);
@@ -128,15 +128,15 @@ internal sealed class Program
     }
 
     /// <summary>
-    /// Run the planner to decide whether to run the GitHub skill function and add the result to the chat history.
+    /// Run the planner to decide whether to run the GitHub plugin function and add the result to the chat history.
     /// </summary>
-    private static async Task<string> PlanGitHubSkillAsync(
-        GitHubSkillOptions gitHubOptions, ActionPlanner planner, OpenAIChatHistory chatHistory, string input, int tokenAllowance, ILogger logger)
+    private static async Task<string> PlanGitHubPluginAsync(
+        GitHubPluginOptions gitHubOptions, ActionPlanner planner, OpenAIChatHistory chatHistory, string input, int tokenAllowance, ILogger logger)
     {
         // Ask the planner to create a plan based off the user's input. If the plan elicits no steps, continue normally.
         Plan plan = await planner.CreatePlanAsync(input);
 
-        // Make sure the plan's state has the GitHub skill configuration values.
+        // Make sure the plan's state has the GitHub plugin configuration values.
         plan.State.Set("repo", gitHubOptions.Repository);
         plan.State.Set("owner", gitHubOptions.Owner);
 
@@ -192,13 +192,13 @@ internal sealed class Program
     }
 
     /// <summary>
-    /// Try to extract json from the planner response as if it were from an OpenAPI skill.
+    /// Try to extract json from the planner response as if it were from an OpenAPI plugin.
     /// </summary>
-    private static bool TryExtractJsonFromOpenApiPlanResult(string openApiSkillResponse, ILogger logger, out string json)
+    private static bool TryExtractJsonFromOpenApiPlanResult(string openApiPluginResponse, ILogger logger, out string json)
     {
         try
         {
-            JsonNode? jsonNode = JsonNode.Parse(openApiSkillResponse);
+            JsonNode? jsonNode = JsonNode.Parse(openApiPluginResponse);
             string contentType = jsonNode?["contentType"]?.ToString() ?? string.Empty;
             if (contentType.StartsWith("application/json", StringComparison.InvariantCultureIgnoreCase))
             {
@@ -212,7 +212,7 @@ internal sealed class Program
         }
         catch (JsonException)
         {
-            logger.LogWarning("Unable to extract JSON from planner response, it is likely not from an OpenAPI skill.");
+            logger.LogWarning("Unable to extract JSON from planner response, it is likely not from an OpenAPI plugin.");
         }
         catch (InvalidOperationException)
         {
