+++ samples/dotnet/OpenApiPluginsExample/RequiredOnPropertyValueAttribute.cs
@@ -4,7 +4,7 @@ using System;
 using System.ComponentModel.DataAnnotations;
 using System.Reflection;
 
-namespace OpenApiSkillsExample;
+namespace OpenApiPluginsExample;
 
 /// <summary>
 /// If the other property is set to the expected value, then this property is required.
