+++ samples/dotnet/OpenApiPluginsExample/appsettings.json
