--- samples/dotnet/graph-api-skills/Properties/launchSettings.json
@@ -1,10 +1,0 @@
-{
-  "profiles": {
-    "GraphSkillsTestApp": {
-      "commandName": "Project",
-      "environmentVariables": {
-        "GraphSkillsTestApp_ClientId": "1234"
-      }
-    }
-  }
-}
